python/langgraph/chained-json-jq-filter/pyproject.toml
[project]
name = "deterministic-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.17",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "mcp>=1.27.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
